Crate sdk_dsn

Crate with DSN shared between sdk farmer and sdk node

Structs

Dsn

Node DSN builder

DsnBuilder

Builder for Dsn.

DsnOptions

Options for DSN

DsnShared

Shared Dsn structure between node and farmer

FarmerPieceCache

Farmer piece cache Piece cache with limited size where pieces closer to provided peer ID are retained.

InConnections

Wrapper with default value for number of incoming connections

ListenAddresses

Wrapper with default value for listen address

NodePieceGetter

Node piece getter (combines DSN and Farmer getters)

OutConnections

Wrapper with default value for number of outgoing connections

PendingInConnections

Wrapper with default value for number of pending incoming connections

PendingOutConnections

Wrapper with default value for number of pending outgoing connections

TargetConnections

Wrapper with default value for number of target connections

Enums

DsnBuilderError

Error type for DsnBuilder

Functions

start_announcements_processor

Start processing announcements received by the network node, returns handle that will stop processing on drop.

Type Definitions

FarmerProviderStorage

Farmer provider storage

NodePieceCache

Node piece cache

NodeProviderStorage

Node provider storage

PieceStore

Farmer piece store

ProviderStorage

General provider storage