Crate sdk_dsn

Crate with DSN shared between sdk farmer and sdk node

Structs

Dsn

Node DSN builder

DsnBuilder

Builder for Dsn.

DsnOptions

Options for DSN

DsnShared

Shared Dsn structure between node and farmer

InConnections

Wrapper with default value for number of incoming connections

ListenAddresses

Wrapper with default value for listen address

OutConnections

Wrapper with default value for number of outgoing connections

PendingInConnections

Wrapper with default value for number of pending incoming connections

PendingOutConnections

Wrapper with default value for number of pending outgoing connections

TargetConnections

Wrapper with default value for number of target connections

Enums

DsnBuilderError

Error type for DsnBuilder

Type Aliases

LocalRecordProvider

A record provider that uses farmer piece cache underneath